+++ Specs/DataSources/SampledPositionPropertySpec.js
@@ -193,7 +193,10 @@ describe("DataSources/SampledPositionProperty", function () {
     var listener = jasmine.createSpy("listener");
     property.definitionChanged.addEventListener(listener);
 
-    var result = property.removeSample(times[1]);
+    var result = property.removeSample(new JulianDate(4, 0));
+    expect(result).toEqual(false);
+
+    result = property.removeSample(times[1]);
 
     expect(listener).toHaveBeenCalledWith(property);